# service/service/employee/_swagger/service/swagger.yaml
openapi: 3.0.3
info:
  version: "1.0"
  title: "Employee Management"
  contact:
    email: "m.ehab.rabea@gmail.com"
    name: "Mohamed Ehab"
servers:
  - url: http://localhost:7002/
...
paths:
...
  /employees/{employeeId}/attendances:
    post:
      tags:
        - EmployeeAttendance
      summary: Create employee attendance record
      operationId: createEmployeeAttendance
      parameters:
        - $ref: '#/components/parameters/PPEmployeeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeAttendanceDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/NewRecordVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
    get:
      tags:
        - EmployeeAttendance
      summary: Retrieve list of employee attendances
      operationId: getAllEmployeeAttendances
      parameters:
        - $ref: '#/components/parameters/PPEmployeeId'
        - $ref: '#/components/parameters/QPEmployeeAttendanceStatus'
        - $ref: '#/components/parameters/QPAttendanceDateFrom'
        - $ref: '#/components/parameters/QPAttendanceDateTo'
        - $ref: '#/components/parameters/QPCheckInFrom'
        - $ref: '#/components/parameters/QPCheckInTo'
        - $ref: '#/components/parameters/QPCheckOutFrom'
        - $ref: '#/components/parameters/QPCheckOutTo'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageNum'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageSize'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderDir'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderBy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeAttendanceResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /employees/attendances:
    get:
      tags:
        - EmployeeAttendance
      summary: Retrieve list of employees attendances
      operationId: getAllEmployeesAttendances
      parameters:
        - $ref: '#/components/parameters/QPEmployeeId'
        - $ref: '#/components/parameters/QPEmployeeAttendanceStatus'
        - $ref: '#/components/parameters/QPAttendanceDateFrom'
        - $ref: '#/components/parameters/QPAttendanceDateTo'
        - $ref: '#/components/parameters/QPCheckInFrom'
        - $ref: '#/components/parameters/QPCheckInTo'
        - $ref: '#/components/parameters/QPCheckOutFrom'
        - $ref: '#/components/parameters/QPCheckOutTo'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageNum'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageSize'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderDir'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderBy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeAttendanceResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /employees/{employeeId}/attendances/{attendanceId}:
    put:
      tags:
        - EmployeeAttendance
      summary: Update employee attendance record
      operationId: updateEmployeeAttendance
      parameters:
        - $ref: '#/components/parameters/PPEmployeeId'
        - $ref: '#/components/parameters/PPAttendanceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeAttendanceDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/NewRecordVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    get:
      tags:
        - EmployeeAttendance
      summary: Retrieve employee attendance record
      operationId: getEmployeeAttendance
      parameters:
        - $ref: '#/components/parameters/PPEmployeeId'
        - $ref: '#/components/parameters/PPAttendanceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeAttendanceVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    delete:
      tags:
        - EmployeeAttendance
      summary: Delete employee attendance record
      operationId: deleteEmployeeAttendance
      parameters:
        - $ref: '#/components/parameters/PPEmployeeId'
        - $ref: '#/components/parameters/PPAttendanceId'
      responses:
        '200':
          description: OK
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /attendances/reports/daily:
    get:
      tags:
        - EmployeeAttendance
      summary: Get daily attendance report
      operationId: getDailyAttendanceReport
      parameters:
        - $ref: '#/components/parameters/QPAttendanceDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyAttendanceReport'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

  # ==================== Trainer Course Endpoints ====================
...
components:
  parameters:
    PPEmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: integer
        format: int32
...
    PPAttendanceId:
      name: attendanceId
      in: path
      required: true
      schema:
        type: integer
        format: int32
...
    QPEmployeeId:
      name: employeeId
      in: query
      required: false
      schema:
        type: integer
        format: int32
...
    QPEmployeeAttendanceStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/EmployeeAttendanceStatus'
    QPAttendanceDateFrom:
      name: attendanceDateFrom
      in: query
      required: false
      schema:
        type: string
        format: date
    QPAttendanceDateTo:
      name: attendanceDateTo
      in: query
      required: false
      schema:
        type: string
        format: date
    QPAttendanceDate:
      name: attendanceDate
      in: query
      required: true
      schema:
        type: string
        format: date
    QPCheckInFrom:
      name: checkInFrom
      in: query
      required: false
      schema:
        type: string
        format: time
    QPCheckInTo:
      name: checkInTo
      in: query
      required: false
      schema:
        type: string
        format: time
    QPCheckOutFrom:
      name: checkOutFrom
      in: query
      required: false
      schema:
        type: string
        format: time
    QPCheckOutTo:
      name: checkOutTo
      in: query
      required: false
      schema:
        type: string
        format: time

    # Course Session Parameters
...
  schemas:
...
    EmployeeAttendanceStatus:
      type: object
      properties:
        id:
          type: integer
          enum: [1, 2, 3, 4]
        code:
          type: string
          enum: ["PRESENT", "ABSENT", "LATE", "EXCUSED"]
        arabicLabel:
          type: string
          enum: ["حاضر", "غائب", "متأخر", "معذور"]
...
    EmployeeAttendanceDTO:
      type: object
      required:
        - attendanceDate
        - status
        - checkInTime
      properties:
        attendanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EmployeeAttendanceStatus'
        checkInTime:
          type: string
          format: time
        checkOutTime:
          type: string
          format: time
        lateTime:
          type: integer
          format: int32
          description: "Late time in minutes"
        note:
          type: string
...
    EmployeeAttendanceVTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        employee:
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupVTO'
        attendanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EmployeeAttendanceStatus'
        checkInTime:
          type: string
          format: time
        checkOutTime:
          type: string
          format: time
        lateTime:
          type: integer
          format: int32
        note:
          type: string
        createdOn:
          type: string
          format: date-time
        createdBy:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
        lastModifiedOn:
          type: string
          format: date-time
        lastModifiedBy:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
...
    EmployeeAttendanceResultSet:
      type: object
      properties:
        total:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeAttendanceListItem'
...
    EmployeeAttendanceListItem:
      type: object
      properties:
        id:
          type: integer
          format: int32
        employee:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
        attendanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EmployeeAttendanceStatus'
        checkInTime:
          type: string
          format: time
        checkOutTime:
          type: string
          format: time
        lateTime:
          type: integer
          format: int32
    DailyAttendanceReport:
      type: object
      properties:
        attendanceDate:
          type: string
          format: date
        totalNumber:
          type: integer
          format: int32
        present:
          type: integer
          format: int32
        absent:
          type: integer
          format: int32
        late:
          type: integer
          format: int32
        excused:
          type: integer
          format: int32
        attendanceRate:
          type: number
          format: double
        details:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeAttendanceListItem'
...