# service/service/employee/employee-core/src/main/java/bs/service/employee/controller/generated/EmployeeAttendanceController.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package bs.service.employee.controller.generated;

import bs.lib.common.model.generated.NewRecordVTO;
import bs.lib.common.model.vto.ErrorVTO;
import bs.lib.sql.db.adapter.model.generated.OrderDirections;
import bs.service.employee.model.enums.EmployeeAttendanceStatus;
import bs.service.employee.model.generated.DailyAttendanceReport;
import bs.service.employee.model.generated.EmployeeAttendanceDTO;
import bs.service.employee.model.generated.EmployeeAttendanceResultSet;
import bs.service.employee.model.generated.EmployeeAttendanceVTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.time.LocalDate;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.annotation.Secured;
import org.springframework.stereotype.Controller;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.6.0")
@Validated
@Controller
@Tag(name = "EmployeeAttendance", description = "the EmployeeAttendance API")
public interface EmployeeAttendanceController {

    /**
     * POST /employees/{employeeId}/attendances : Create employee attendance record
     *
     * @param employeeId
     *            (required)
     * @param employeeAttendanceDTO
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "createEmployeeAttendance", summary = "Create employee attendance record", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = NewRecordVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.POST, value = "/employees/{employeeId}/attendances", produces = {
            "application/json" }, consumes = { "application/json" })

    ResponseEntity<NewRecordVTO> _createEmployeeAttendance(
            @Parameter(name = "employeeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("employeeId") Integer employeeId,
            @Parameter(name = "EmployeeAttendanceDTO", description = "", required = true) @Valid @RequestBody EmployeeAttendanceDTO employeeAttendanceDTO);

    /**
     * DELETE /employees/{employeeId}/attendances/{attendanceId} : Delete employee attendance record
     *
     * @param employeeId
     *            (required)
     * @param attendanceId
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "deleteEmployeeAttendance", summary = "Delete employee attendance record", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK"),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.DELETE, value = "/employees/{employeeId}/attendances/{attendanceId}", produces = {
            "application/json" })

    ResponseEntity<Void> _deleteEmployeeAttendance(
            @Parameter(name = "employeeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("employeeId") Integer employeeId,
            @Parameter(name = "attendanceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("attendanceId") Integer attendanceId);

    /**
     * GET /employees/{employeeId}/attendances : Retrieve list of employee attendances
     *
     * @param employeeId
     *            (required)
     * @param status
     *            (optional)
     * @param attendanceDateFrom
     *            (optional)
     * @param attendanceDateTo
     *            (optional)
     * @param checkInFrom
     *            (optional)
     * @param checkInTo
     *            (optional)
     * @param checkOutFrom
     *            (optional)
     * @param checkOutTo
     *            (optional)
     * @param pageNum
     *            (optional, default to 0)
     * @param pageSize
     *            (optional, default to 25)
     * @param orderDir
     *            Order Direction (optional)
     * @param orderBy
     *            Order By Attribute (optional)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getAllEmployeeAttendances", summary = "Retrieve list of employee attendances", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = EmployeeAttendanceResultSet.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/employees/{employeeId}/attendances", produces = {
            "application/json" })

    ResponseEntity<EmployeeAttendanceResultSet> _getAllEmployeeAttendances(
            @Parameter(name = "employeeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("employeeId") Integer employeeId,
            @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid EmployeeAttendanceStatus status,
            @Parameter(name = "attendanceDateFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "attendanceDateFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate attendanceDateFrom,
            @Parameter(name = "attendanceDateTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "attendanceDateTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate attendanceDateTo,
            @Parameter(name = "checkInFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkInFrom", required = false) String checkInFrom,
            @Parameter(name = "checkInTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkInTo", required = false) String checkInTo,
            @Parameter(name = "checkOutFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkOutFrom", required = false) String checkOutFrom,
            @Parameter(name = "checkOutTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkOutTo", required = false) String checkOutTo,
            @Min(0) @Parameter(name = "pageNum", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
            @Min(1) @Max(100) @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
            @Parameter(name = "orderDir", description = "Order Direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderDir", required = false) OrderDirections orderDir,
            @Parameter(name = "orderBy", description = "Order By Attribute", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderBy", required = false) String orderBy);

    /**
     * GET /employees/attendances : Retrieve list of employees attendances
     *
     * @param employeeId
     *            (optional)
     * @param status
     *            (optional)
     * @param attendanceDateFrom
     *            (optional)
     * @param attendanceDateTo
     *            (optional)
     * @param checkInFrom
     *            (optional)
     * @param checkInTo
     *            (optional)
     * @param checkOutFrom
     *            (optional)
     * @param checkOutTo
     *            (optional)
     * @param pageNum
     *            (optional, default to 0)
     * @param pageSize
     *            (optional, default to 25)
     * @param orderDir
     *            Order Direction (optional)
     * @param orderBy
     *            Order By Attribute (optional)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getAllEmployeesAttendances", summary = "Retrieve list of employees attendances", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = EmployeeAttendanceResultSet.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/employees/attendances", produces = { "application/json" })

    ResponseEntity<EmployeeAttendanceResultSet> _getAllEmployeesAttendances(
            @Parameter(name = "employeeId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "employeeId", required = false) Integer employeeId,
            @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid EmployeeAttendanceStatus status,
            @Parameter(name = "attendanceDateFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "attendanceDateFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate attendanceDateFrom,
            @Parameter(name = "attendanceDateTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "attendanceDateTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate attendanceDateTo,
            @Parameter(name = "checkInFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkInFrom", required = false) String checkInFrom,
            @Parameter(name = "checkInTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkInTo", required = false) String checkInTo,
            @Parameter(name = "checkOutFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkOutFrom", required = false) String checkOutFrom,
            @Parameter(name = "checkOutTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "checkOutTo", required = false) String checkOutTo,
            @Min(0) @Parameter(name = "pageNum", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
            @Min(1) @Max(100) @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
            @Parameter(name = "orderDir", description = "Order Direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderDir", required = false) OrderDirections orderDir,
            @Parameter(name = "orderBy", description = "Order By Attribute", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderBy", required = false) String orderBy);

    /**
     * GET /attendances/reports/daily : Get daily attendance report
     *
     * @param attendanceDate
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getDailyAttendanceReport", summary = "Get daily attendance report", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = DailyAttendanceReport.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/attendances/reports/daily", produces = { "application/json" })

    ResponseEntity<DailyAttendanceReport> _getDailyAttendanceReport(
            @NotNull @Parameter(name = "attendanceDate", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "attendanceDate", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate attendanceDate);

    /**
     * GET /employees/{employeeId}/attendances/{attendanceId} : Retrieve employee attendance record
     *
     * @param employeeId
     *            (required)
     * @param attendanceId
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getEmployeeAttendance", summary = "Retrieve employee attendance record", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = EmployeeAttendanceVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/employees/{employeeId}/attendances/{attendanceId}", produces = {
            "application/json" })

    ResponseEntity<EmployeeAttendanceVTO> _getEmployeeAttendance(
            @Parameter(name = "employeeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("employeeId") Integer employeeId,
            @Parameter(name = "attendanceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("attendanceId") Integer attendanceId);

    /**
     * PUT /employees/{employeeId}/attendances/{attendanceId} : Update employee attendance record
     *
     * @param employeeId
     *            (required)
     * @param attendanceId
     *            (required)
     * @param employeeAttendanceDTO
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "updateEmployeeAttendance", summary = "Update employee attendance record", tags = {
            "EmployeeAttendance" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = NewRecordVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.PUT, value = "/employees/{employeeId}/attendances/{attendanceId}", produces = {
            "application/json" }, consumes = { "application/json" })

    ResponseEntity<NewRecordVTO> _updateEmployeeAttendance(
            @Parameter(name = "employeeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("employeeId") Integer employeeId,
            @Parameter(name = "attendanceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("attendanceId") Integer attendanceId,
            @Parameter(name = "EmployeeAttendanceDTO", description = "", required = true) @Valid @RequestBody EmployeeAttendanceDTO employeeAttendanceDTO);

}